requirements.txt
beautifulsoup4==4.12.3
packaging==24.1
pillow==10.4.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
soupsieve==2.6
